refactor: move chat history into llm agent interfaces
