+++ src/xtalk/llm_agent/__init__.py
@@ -1,5 +1,5 @@
 from .tools.utils import ToolCallResultArgs, ToolCallResult
-from .interfaces import Agent, AgentContext, AgentOutput
+from .interfaces import Agent, AgentContext, AgentOutput, ChatHistory, PlaybackAIMessageMeta
 from .dummy import DummyAgent
 from .default import DefaultAgent
 from .lts import LTSAgent
@@ -9,10 +9,12 @@ __all__ = [
     "Agent",
     "AgentContext",
     "AgentOutput",
+    "ChatHistory",
     "DummyAgent",
     "DefaultAgent",
     "LTSAgent",
     "ExperimentalAgent",
+    "PlaybackAIMessageMeta",
     "ToolCallResultArgs",
     "ToolCallResult",
 ]
