+++ src/xtalk/llm_agent/interfaces.py
@@ -1,8 +1,9 @@
 import asyncio
 from abc import ABC, abstractmethod
+from dataclasses import dataclass
 from typing import Iterable, TypedDict, AsyncIterator, Callable, Any, Union, TypeVar
 
-from langchain_core.messages import ToolCall
+from langchain_core.messages import BaseMessage, SystemMessage, AIMessage, ToolCall
 from langchain_core.tools import BaseTool
 
 from .tools.utils import ToolCallResult
@@ -23,6 +24,118 @@ class AgentContext(TypedDict):
 
 AgentOutput = Union[str, ToolCall, ToolCallResult]
 T = TypeVar("T")
+
+
+@dataclass
+class PlaybackAIMessageMeta:
+    """Track merge state for one playback-managed assistant message."""
+
+    final: bool = False
+    prefix: str | None = None
+
+
+class ChatHistory:
+    """Manage chat history plus playback-aware assistant-message merging."""
+
+    def __init__(self, system_prompt: str) -> None:
+        """Initialize the history with one system message.
+
+        Parameters
+        ----------
+        system_prompt:
+            The system prompt to place at the start of the message list.
+        """
+
+        self._messages: list[BaseMessage] = [SystemMessage(content=system_prompt)]
+        self._playback_ai_meta: dict[int, PlaybackAIMessageMeta] = {}
+
+    @property
+    def messages(self) -> list[BaseMessage]:
+        """Return the current chat-history message list."""
+
+        return self._messages
+
+    def append_message(self, message: BaseMessage) -> None:
+        """Append one message to the history unchanged.
+
+        Parameters
+        ----------
+        message:
+            The message to append.
+        """
+
+        self._messages.append(message)
+
+    def append_or_update_ai_message(self, full_text: str, *, final: bool) -> None:
+        """Append or merge one playback-managed assistant message.
+
+        Parameters
+        ----------
+        full_text:
+            The cumulative assistant text confirmed by playback.
+        final:
+            Whether this update closes the playback-managed assistant message.
+        """
+
+        if not full_text:
+            return
+
+        previous_playback = self._find_previous_playback_ai_message()
+        if previous_playback is not None:
+            index, message, meta = previous_playback
+            if index == len(self._messages) - 1 and not meta.final:
+                if meta.prefix and full_text.startswith(meta.prefix):
+                    message.content = full_text[len(meta.prefix) :]
+                else:
+                    message.content = full_text
+                    meta.prefix = None
+                meta.final = final
+                return
+
+            previous_full_text = self._resolve_playback_full_text(message, meta)
+            if not meta.final and full_text.startswith(previous_full_text):
+                suffix = full_text[len(previous_full_text) :]
+                meta.final = final
+                if not suffix:
+                    return
+                self._messages.append(AIMessage(content=suffix))
+                self._playback_ai_meta[len(self._messages) - 1] = PlaybackAIMessageMeta(
+                    final=final,
+                    prefix=previous_full_text,
+                )
+                return
+
+        self._append_playback_ai_message(full_text, final=final)
+
+    def _append_playback_ai_message(self, content: str, *, final: bool) -> int:
+        """Append one playback-managed assistant message."""
+
+        message_index = len(self._messages)
+        self.append_message(AIMessage(content=content))
+        self._playback_ai_meta[message_index] = PlaybackAIMessageMeta(final=final)
+        return message_index
+
+    def _find_previous_playback_ai_message(
+        self,
+    ) -> tuple[int, AIMessage, PlaybackAIMessageMeta] | None:
+        """Return the latest playback-managed assistant message if present."""
+
+        for index in range(len(self._messages) - 1, -1, -1):
+            if index not in self._playback_ai_meta:
+                continue
+            message = self._messages[index]
+            if isinstance(message, AIMessage):
+                return index, message, self._playback_ai_meta[index]
+        return None
+
+    @staticmethod
+    def _resolve_playback_full_text(
+        message: AIMessage,
+        meta: PlaybackAIMessageMeta,
+    ) -> str:
+        """Resolve the full playback text represented by one assistant message."""
+
+        return f"{meta.prefix or ''}{message.content}"
 
 
 class Agent(ABC):
